Added xds sync test
