+++ performance/tests/test.env
@@ -4,3 +4,4 @@ export APPNSPREFIX=app-perf-test
 export GODEBUG="x509ignoreCN=0"
 export NSCOUNTBUNDLE="5"
 export NSACCEPTANCETIME=3
+export XDSPUSHACCEPTANCETIME=1
