+++ pkg/performance/yaml_vars.go
@@ -19,12 +19,14 @@ import (
 )
 
 var (
-	smcpName         string            = util.Getenv("SMCPNAME", "basic")
-	meshNamespace    string            = util.Getenv("MESHNAMESPACE", "istio-system")
-	appNSPrefix      string            = util.Getenv("APPNSPREFIX", "app-perf-test")
-	nsCountBundle    string            = util.Getenv("NSCOUNTBUNDLE", "10,50,100,500")
-	nsAcceptanceTime string            = util.Getenv("NSACCEPTANCETIME", "5")
-	prometheusAPIMap map[string]string = map[string]string{
-		"xds_ppctb": "histogram_quantile(0.99, sum(rate(pilot_proxy_convergence_time_bucket[1m])) by (le))",
+	smcpName              string            = util.Getenv("SMCPNAME", "basic")
+	meshNamespace         string            = util.Getenv("MESHNAMESPACE", "istio-system")
+	appNSPrefix           string            = util.Getenv("APPNSPREFIX", "app-perf-test")
+	nsCountBundle         string            = util.Getenv("NSCOUNTBUNDLE", "10,50,100,500")
+	nsAcceptanceTime      string            = util.Getenv("NSACCEPTANCETIME", "5")
+	xdsPushAcceptanceTime string            = util.Getenv("XDSPUSHACCEPTANCETIME", "1")
+	prometheusAPIMap      map[string]string = map[string]string{
+		"xds_ppctc": "pilot_proxy_convergence_time_count",
+		"xds_ppctb": "pilot_proxy_convergence_time_bucket{le=\"" + xdsPushAcceptanceTime + "\"}",
 	}
 )
